requirements.txt
uvicorn
rdflib
SPARQLWrapper
pydantic
fastapi
pydantic-settings
python-dotenv
requests